feat(test-utils): move the toEmitValue/s jest matchers to test-utils and add test script
