+++ packages/grafana-test-utils/jest-setup.js
@@ -1,0 +1,3 @@
+import { matchers } from '@grafana/test-utils/matchers';
+
+expect.extend(matchers);
