+++ public/test/setupTests.ts
@@ -7,9 +7,9 @@ import i18next from 'i18next';
 import failOnConsole from 'jest-fail-on-console';
 import { initReactI18next } from 'react-i18next';
 
-import getEnvConfig from '../../scripts/webpack/env-util';
+import { matchers } from '@grafana/test-utils/matchers';
 
-import { matchers } from './matchers';
+import getEnvConfig from '../../scripts/webpack/env-util';
 
 const config = getEnvConfig() as Record<string, string | boolean>;
